+++ leitor-instancias/src/solucao.cpp
@@ -4,7 +4,7 @@
 
 void exibirSolucao(Solucao &s)
 {
-    for(int i = 0; i < s.sequencia.size() - 1; i++)
+    for(size_t i = 0; i < s.sequencia.size() - 1; i++)
         std::cout << s.sequencia[i] << " -> ";
     std::cout << s.sequencia.back() << std::endl;
 }
@@ -12,6 +12,6 @@ void exibirSolucao(Solucao &s)
 void calcularValorObj (Solucao &s, double **matrizAdj)
 {
     s.valorObj = 0;
-    for(int i = 0; i < s.sequencia.size() - 1; i++)
+    for(size_t i = 0; i < s.sequencia.size() - 1; i++)
     s.valorObj += matrizAdj[s.sequencia[i] - 1][s.sequencia[i+1] - 1];
 }
